+++ src/App.tsx
@@ -1,7 +1,5 @@
-import { ThemeProvider } from 'styled-components'
-import { theme } from './styles/theme'
-import GlobalStyles from './styles/GlobalStyles'
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
+
 import Home from './pages/Home/Home'
 import Login from './pages/LoginAndSignup/Login'
 import Signup from './pages/LoginAndSignup/Signup'
@@ -11,6 +9,11 @@ import AddItem from './pages/AddItem/AddItem'
 import Privacy from './pages/Privacy/privacy’'
 import Faq from './pages/Faq/Faq'
 import NavVArLayout from './Layout/NavVArLayout'
+
+import { ThemeProvider } from 'styled-components'
+import { theme } from './styles/theme'
+import GlobalStyles from './styles/GlobalStyles'
+
 function App() {
   return (
     <ThemeProvider theme={theme}>
