--- src/api/client/interceptors.jsx
@@ -1,8 +1,0 @@
-export const requestInterceptor = (config) => {
-  const { token } = localStorage.getItem('token') ?? ''
-
-  if (token) {
-    config.headers.set('Authorization', `Basic${token}`)
-  }
-  return config
-}
