+++ src/api/services/productService.tsx
@@ -1,23 +1,23 @@
 import { Axios, AxiosResponse } from 'axios'
 import requestor from '../client/requestor'
-import { Product } from '../../types/product'
+import { GetProductIdTypes } from '../../types/product'
 
 class ProductService {
   //틀릴 수도
-  postProduct(body: Product) {
+  postProduct(body: GetProductIdTypes) {
     const requestBody = {
       ...body,
     }
     return requestor.post(`/products`, requestBody)
   }
 
-  getProduct(page, pageSize, orderBy, keyword) {
+  getProduct(page: number, pageSize: number, orderBy: string, keyword: string) {
     return requestor.get(
       `/products?page=${page}&pageSize=${pageSize}&orderBy=${orderBy}&keyword=${keyword}`
     )
   }
 
-  getProductId(productId: number): Promise<AxiosResponse<Product>> {
+  getProductId(productId: number): Promise<AxiosResponse<GetProductIdTypes>> {
     return requestor.get(`/products/${productId}`)
   }
