+++ src/pages/Items/Items.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { useNavigate, useLocation } from 'react-router-dom'
 
+import { GetProductIdTypes } from '../../types/product'
 import BestItems from './BestItems'
 import RecentItems from './RecentItems'
 import DropDown from '../../component/common/DropDown'
@@ -14,6 +15,162 @@ import ArrowRight from '../../assets/svg/ArrowRight.svg'
 import styled, { css } from 'styled-components'
 import { theme } from '../../styles/theme'
 import { textStyle } from '../../styles/textStyle'
+
+type SelectOption = {
+  value: string
+  name: string
+}
+
+const Items = () => {
+  const navigate = useNavigate()
+  const location = useLocation() // location을 써서 url 주소를 가져와 선택된 옵션에 맞게 변경경
+
+  const [bestProducts, setBestProducts] = useState<GetProductIdTypes[]>([])
+  const [sortedProducts, setSortedProducts] = useState<GetProductIdTypes[]>([])
+  const [totalItems, setTotalItems] = useState(0) // 전체 상품 개수 저장
+
+  const selectList: SelectOption[] = [
+    { value: 'recent', name: '최신순' },
+    { value: 'favorite', name: '좋아요순' },
+  ]
+  const [selectedOption, setSelectedOption] = useState(selectList[0].value)
+  const itemsPerPage = 10 // 페이지 네이션
+  const [currentPage, setCurrentPage] = useState(1)
+
+  useEffect(() => {
+    const queryParams = new URLSearchParams(location.search)
+    const orderBy = queryParams.get('orderBy') ?? 'recent'
+    const page = parseInt(queryParams.get('page') ?? '1', 10)
+
+    setSelectedOption(orderBy)
+    setCurrentPage(page)
+  }, [location.search])
+
+  const updateURL = (orderBy: string, page: number) => {
+    navigate(`?orderBy=${orderBy}&page=${page}`)
+  }
+
+  const handleAdditem = () => {
+    navigate(`/additem`)
+  }
+  // BestItems 데이터 불러오기, bestProducts
+  useEffect(() => {
+    productService.getProduct(1, 10, 'favorite', '').then((response) => {
+      setBestProducts(response.data)
+    })
+  }, [])
+
+  // 페이지네이션
+  useEffect(() => {
+    productService
+      .getProduct(currentPage, 10, selectedOption, '')
+      .then((response) => {
+        const sorted = [...(response.data.list || [])].sort((a, b) => {
+          if (selectedOption === 'recent') {
+            return (
+              new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
+            ) // Date 쓸 때는 getTime을 써야 함함
+          } else if (selectedOption === 'favorite') {
+            return b.favoriteCount - a.favoriteCount
+          }
+          return 0
+        })
+        console.log(sorted)
+        setSortedProducts(sorted)
+        setTotalItems(response.data.totalCount)
+      })
+  }, [currentPage, selectedOption])
+
+  const totalPages = Math.ceil(totalItems / itemsPerPage)
+
+  const pages = (() => {
+    if (totalPages <= 5) {
+      return Array.from({ length: totalPages }, (_, i) => i + 1)
+    }
+    if (currentPage <= 3) {
+      return [1, 2, 3, 4, 5]
+    }
+    if (currentPage >= totalPages - 2) {
+      return Array.from({ length: 5 }, (_, i) => totalPages - 4 + i)
+    }
+    return Array.from({ length: 5 }, (_, i) => currentPage - 2 + i)
+  })()
+
+  const handleNextPage = () => {
+    setCurrentPage((prev) => {
+      const nextPage = prev + 1
+      updateURL(selectedOption, nextPage)
+      return nextPage
+    })
+  }
+
+  const handlePrevPage = () => {
+    setCurrentPage((prev) => {
+      const prevPage = prev > 1 ? prev - 1 : 1
+      updateURL(selectedOption, prevPage)
+      return prevPage
+    })
+  }
+
+  return (
+    <>
+      <Bone>
+        <BestItems products={bestProducts} />
+
+        <NavVAr>
+          <NavTitle>전체상품</NavTitle>
+          <NavRightWrapper>
+            <SearchIcon src={Search} alt={`검색 아이콘`} />
+            <NavSearch placeholder="검색할 상품 입력해주세요" />
+            <ButtonWrapper>
+              <Button
+                size={42.5}
+                onClick={handleAdditem}
+                paddingHeight={8}
+                paddingWidth={23}
+              >
+                상품 등록하기
+              </Button>
+            </ButtonWrapper>
+            <DropDown
+              selectList={selectList}
+              selected={selectedOption}
+              onChange={(value) => {
+                setSelectedOption(value)
+                updateURL(value, currentPage)
+              }}
+            />
+          </NavRightWrapper>
+        </NavVAr>
+
+        <RecentItems products={sortedProducts} />
+
+        <Pagenation>
+          <LeftButton onClick={handlePrevPage}>
+            <img src={ArrowLeft} />
+          </LeftButton>
+          {pages.map((page) => (
+            <PageButton
+              key={page}
+              $isActive={currentPage === page}
+              onClick={() => {
+                setCurrentPage(page)
+                updateURL(selectedOption, page)
+              }}
+            >
+              {page}
+            </PageButton>
+          ))}
+          <LeftButton onClick={handleNextPage}>
+            <img src={ArrowRight} />
+          </LeftButton>
+        </Pagenation>
+      </Bone>
+    </>
+  )
+}
+
+export default Items
 
 const Bone = styled.div`
   width: 75rem;
@@ -142,7 +299,10 @@ const LeftButton = styled.button`
   }
 `
 
-const PageButton = styled.button`
+interface PageButtonProps {
+  $isActive: boolean
+}
+const PageButton = styled.button<PageButtonProps>`
   width: 40px;
   height: 40px;
   border-radius: 50%;
@@ -158,154 +318,3 @@ const PageButton = styled.button`
       color: ${theme.colors.SecondaryGray[50]};
     `}
 `
-
-const Items = () => {
-  const navigate = useNavigate()
-  const location = useLocation() // location을 써서 url 주소를 가져와 선택된 옵션에 맞게 변경경
-
-  const [bestProducts, setBestProducts] = useState([])
-  const [sortedProducts, setSortedProducts] = useState([])
-  const [totalItems, setTotalItems] = useState(0) // 전체 상품 개수 저장
-
-  const selectList = [
-    { value: 'recent', name: '최신순' },
-    { value: 'favorite', name: '좋아요순' },
-  ]
-
-  const [selectedOption, setSelectedOption] = useState(selectList[0].value)
-  const itemsPerPage = 10 // 페이지 네이션
-  const [currentPage, setCurrentPage] = useState(1)
-
-  useEffect(() => {
-    const queryParams = new URLSearchParams(location.search)
-    const orderBy = queryParams.get('orderBy') || 'recent'
-    const page = parseInt(queryParams.get('page'), 10) || 1
-
-    setSelectedOption(orderBy)
-    setCurrentPage(page)
-  }, [location.search])
-
-  const updateURL = (orderBy, page) => {
-    navigate(`?orderBy=${orderBy}&page=${page}`)
-  }
-
-  const handleAdditem = () => {
-    navigate(`/additem`)
-  }
-  // BestItems 데이터 불러오기, bestProducts
-  useEffect(() => {
-    productService.getProduct(1, 10, 'favorite', '').then((response) => {
-      setBestProducts(response.data)
-    })
-  }, [])
-
-  // 페이지네이션
-  useEffect(() => {
-    productService
-      .getProduct(currentPage, 10, selectedOption, '')
-      .then((response) => {
-        const sorted = [...(response.data.list || [])].sort((a, b) => {
-          if (selectedOption === 'recent') {
-            return new Date(b.createdAt) - new Date(a.createdAt)
-          } else if (selectedOption === 'favorite') {
-            return b.favoriteCount - a.favoriteCount
-          }
-          return 0
-        })
-        console.log(sorted)
-        setSortedProducts(sorted)
-        setTotalItems(response.data.totalCount)
-      })
-  }, [currentPage, selectedOption])
-
-  const totalPages = Math.ceil(totalItems / itemsPerPage)
-
-  const pages = (() => {
-    if (totalPages <= 5) {
-      return Array.from({ length: totalPages }, (_, i) => i + 1)
-    }
-    if (currentPage <= 3) {
-      return [1, 2, 3, 4, 5]
-    }
-    if (currentPage >= totalPages - 2) {
-      return Array.from({ length: 5 }, (_, i) => totalPages - 4 + i)
-    }
-    return Array.from({ length: 5 }, (_, i) => currentPage - 2 + i)
-  })()
-
-  const handleNextPage = () => {
-    setCurrentPage((prev) => {
-      const nextPage = prev + 1
-      updateURL(selectedOption, nextPage)
-      return nextPage
-    })
-  }
-
-  const handlePrevPage = () => {
-    setCurrentPage((prev) => {
-      const prevPage = prev > 1 ? prev - 1 : 1
-      updateURL(selectedOption, prevPage)
-      return prevPage
-    })
-  }
-
-  return (
-    <>
-      <Bone>
-        <BestItems products={bestProducts} />
-
-        <NavVAr>
-          <NavTitle>전체상품</NavTitle>
-          <NavRightWrapper>
-            <SearchIcon src={Search} alt={`검색 아이콘`} />
-            <NavSearch placeholder="검색할 상품 입력해주세요" />
-            <ButtonWrapper>
-              <Button
-                variant="primary"
-                size={42.5}
-                onClick={handleAdditem}
-                paddingHeight={8}
-                paddingWidth={23}
-              >
-                상품 등록하기
-              </Button>
-            </ButtonWrapper>
-            <DropDown
-              selectList={selectList}
-              selected={selectedOption}
-              onChange={(value) => {
-                setSelectedOption(value)
-                updateURL(value, currentPage)
-              }}
-            />
-          </NavRightWrapper>
-        </NavVAr>
-
-        <RecentItems products={sortedProducts} />
-
-        <Pagenation>
-          <LeftButton onClick={handlePrevPage}>
-            <img src={ArrowLeft} />
-          </LeftButton>
-          {pages.map((page) => (
-            <PageButton
-              key={page}
-              $isActive={currentPage === page}
-              onClick={() => {
-                setCurrentPage(page)
-                updateURL(selectedOption, page)
-              }}
-            >
-              {page}
-            </PageButton>
-          ))}
-          <LeftButton onClick={handleNextPage}>
-            <img src={ArrowRight} />
-          </LeftButton>
-        </Pagenation>
-      </Bone>
-    </>
-  )
-}
-
-export default Items
